Warning fixes:
* Corrected some CUDA version include guards in `graph/instance.hpp` which were enabling an unused parameter in CUDA 11.0 - 11.3.x
* Explicitly cast a `size_t` to `unsigned int` in `module.hpp`, to avoid narrowing warnings
* In the asyncAPI example, now Using iostreams printing rather than a `printf()` with an inexact format specifier (`size_t` vs `unsigned long`)
* Now using float literals, rather than double literals, to set float variables or fill float buffers, in the bandwidthTest and jacobiCudaGraphs examples
* streamOrderedAllocation example: Explicit cast from `size_t` to `int` to avoid a warning about narrowing

graph_unused

+++ examples/modified_cuda_samples/streamOrderedAllocation/streamOrderedAllocation.cu
@@ -115,7 +115,7 @@ int basicStreamOrderedAllocation(
 	auto d_c = span<float>(stream.enqueue.allocate(c.size() * sizeof(float)));
 	stream.enqueue.copy(d_a, a);
 	stream.enqueue.copy(d_b, b);
-	stream.enqueue.kernel_launch(vectorAddGPU, launch_config, d_a.data(), d_b.data(), d_c.data(), c.size());
+	stream.enqueue.kernel_launch(vectorAddGPU, launch_config, d_a.data(), d_b.data(), d_c.data(), (int) c.size());
 	stream.enqueue.free(d_a);
 	stream.enqueue.free(d_b);
 	stream.enqueue.copy(c, d_c);
